pyproject.toml
[project]
name = "trifold"
dynamic = ["version"]
description = "Sample project with React and FastAPI on Databricks Apps"
readme = "README.md"
authors = [{ name = "renardeinside", email = "polarpersonal@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.58.0",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
]

[tool.uv-dynamic-versioning]
format = "v{base}+{commit}.{timestamp}"

[tool.hatch.build.hooks.custom]
path = "hooks/app_build.py"
